requirements.txt
praw>=7.7.0
pandas>=2.0.0
python-dotenv>=1.0.0
pyyaml>=6.0
tqdm>=4.65.0
matplotlib>=3.7.0
seaborn>=0.13.0
textblob>=0.18.0
